+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 ## 0.0.2
 
 - Add disable property for button
+
+
+## 0.0.3
+
+- Remove required from onTap button
+- changes default property's values